requirements.txt
Django==6.0.4
django-allauth[socialaccount]==65.16.1
